chore: Update karma.conf.ci.js to include browser timeout configurations
The karma.conf.ci.js file was modified to include browser timeout configurations. The browserNoActivityTimeout was set to 1 minute, browserDisconnectTimeout was set to 10 seconds, browserDisconnectTolerance was set to 3, and captureTimeout was set to 2 minutes. These changes ensure that the tests run smoothly and handle browser disconnections and timeouts appropriately.

Note: This commit message follows the established convention of using the `chore` type for changes related to build processes, tooling, and configuration.